Add statcast_pitcher_spin method with testing
statcast_pitcher_spin extends statcast_pitcher by adding spin data
back into the file, replacing the deprecated spin columns. The
math and physics behind the calculations were modeled off of
Professor Alan Nathan's work at the University of Illinois. I have
no impression of what information was in the original spin columns
before they were deprecated, but they now have the magnitude of
movement cause by spin in the X and Z directions ('Mx' and 'Mz')
as well as the axis of rotation ('phi').

statcast_pitcher_spin was developed with the test driven development
method, so a testing folder containing the testing file and data
was added as well.

Documentation is next on the to-dos for this method.